Providerと処理を分離させる
