+++ frontend/src/components/pages/HomeWrapper.tsx
@@ -2,10 +2,10 @@ import { useContext } from 'react'
 import { Navigate } from 'react-router-dom'
 
 import { WelcomeWrapper } from 'components/pages/WelcomeWrapper'
-import { AuthContext } from 'context/AuthContext'
+import { UserContext } from 'context/UserContext'
 
 export const HomeWrapper = () => {
-  const { isLoaded, currentUser, currentFirebaseUser } = useContext(AuthContext)
+  const { isLoaded, currentUser, currentFirebaseUser } = useContext(UserContext)
   if (isLoaded && currentUser && currentFirebaseUser) {
     return <Navigate to="/home" />
   }
