+++ frontend/src/providers/PaymentProvider.tsx
@@ -1,17 +1,11 @@
-import { useContext, useEffect, useMemo, useState } from 'react'
+import { useState } from 'react'
 
-import { AuthContext } from 'context/AuthContext'
 import { PaymentContext } from 'context/PaymentContext'
-import { getPayments } from 'lib/api/payment'
-import { getTeamStatus } from 'lib/api/team'
-import { auth } from 'lib/firebase'
-import { useToast } from 'lib/toast'
 
 import type { PaymentListGroupByPaidAt } from 'types/paymentListGroupByPaidAt'
 import type { TeamStatus } from 'types/teamStatus'
 
 export const PaymentProvider = ({ children }: { children: React.ReactElement }) => {
-  const { currentUser } = useContext(AuthContext)
   const [paymentList, setPaymentList] = useState<PaymentListGroupByPaidAt[]>([])
   const [updatePaymentList, setUpdatePaymentList] = useState<boolean>(false)
   const [teamStatus, setTeamStatus] = useState<TeamStatus>({
@@ -23,64 +17,20 @@ export const PaymentProvider = ({ children }: { children: React.ReactElement }) 
   })
   const [isPaymentListLoaded, setIsPaymentListLoaded] = useState<boolean>(false)
   const [isTeamStatusLoaded, setIsTeamStatusLoaded] = useState<boolean>(false)
-  const { errorToast } = useToast()
-  const value = useMemo(
-    () => ({
-      paymentList,
-      setPaymentList,
-      isPaymentListLoaded,
-      setIsPaymentListLoaded,
-      teamStatus,
-      setTeamStatus,
-      isTeamStatusLoaded,
-      setIsTeamStatusLoaded,
-      updatePaymentList,
-      setUpdatePaymentList,
-    }),
-    [paymentList, isPaymentListLoaded, teamStatus, isTeamStatusLoaded, updatePaymentList]
-  )
 
-  const handleGetPayments = async () => {
-    setIsPaymentListLoaded(false)
-    const idToken = await auth.currentUser?.getIdToken(true)
-    try {
-      const res = await getPayments(idToken)
-      setPaymentList(res?.data)
-    } catch {
-      errorToast('支払情報の取得ができませんでした', '時間をおいてから再読み込みをしてください')
-    } finally {
-      setIsPaymentListLoaded(true)
-    }
+  // eslint-disable-next-line react/jsx-no-constructed-context-values
+  const value = {
+    paymentList,
+    setPaymentList,
+    isPaymentListLoaded,
+    setIsPaymentListLoaded,
+    teamStatus,
+    setTeamStatus,
+    isTeamStatusLoaded,
+    setIsTeamStatusLoaded,
+    updatePaymentList,
+    setUpdatePaymentList,
   }
-
-  const handleGetTeamStatus = async () => {
-    setIsTeamStatusLoaded(false)
-    const idToken = await auth.currentUser?.getIdToken(true)
-    if (currentUser) {
-      try {
-        const res = await getTeamStatus(currentUser.teamId, idToken)
-        setTeamStatus(res?.data)
-      } catch {
-        errorToast('情報の取得に失敗しました', '時間をおいてから再読み込みをしてください')
-      } finally {
-        setIsTeamStatusLoaded(true)
-      }
-    }
-  }
-
-  useEffect(() => {
-    handleGetPayments().catch((err) => {
-      console.log(err)
-    })
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [updatePaymentList, currentUser])
-
-  useEffect(() => {
-    handleGetTeamStatus().catch((err) => {
-      console.log(err)
-    })
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [paymentList])
 
   return <PaymentContext.Provider value={value}>{children}</PaymentContext.Provider>
 }
